Modif pr, edit plus fonctionnel
