+++ src/main/core/Resources/styles/administration/parameters.less
@@ -95,6 +95,10 @@
     margin-bottom: 10px;
 }
 
+.color-chart-actions {
+    margin-left: auto;
+}
+
 .color-chart-title {
     font-size: 14px;
     font-weight: 600;
@@ -107,13 +111,13 @@
     margin-right: 10px;
 }
 
-.color-colors-list {
+.color-chart-colors-list {
     display: flex;
     flex: 1 1 auto;
     justify-content: left;
 }
 
-.color-pastille {
+.color-chart-dot {
     border-radius: 50%;
     width: 25px;
     height: 25px;
