--- src/main/theme/Resources/modules/administration/appearance/modals/color-chart-creation/store/index.js
@@ -1,9 +1,0 @@
-import {reducer} from '#/main/theme/administration/appearance/modals/color-chart-creation/store/reducer'
-import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-creation/store/selectors'
-import {actions} from '#/main/theme/administration/appearance/modals/color-chart-creation/store/actions'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
