--- src/main/theme/Resources/modules/administration/appearance/modals/color-chart-creation/store/selectors.js
@@ -1,5 +1,0 @@
-const STORE_NAME = 'colorChartCreation'
-
-export const selectors = {
-  STORE_NAME
-}
