--- src/main/theme/Resources/modules/administration/appearance/modals/color-chart-edit/store/index.js
@@ -1,9 +1,0 @@
-import {reducer} from '#/main/theme/administration/appearance/modals/color-chart-edit/store/reducer'
-import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-edit/store/selectors'
-import {actions} from '#/main/theme/administration/appearance/modals/color-chart-edit/store/actions'
-
-export {
-  actions,
-  reducer,
-  selectors
-}
