--- src/main/theme/Resources/modules/administration/appearance/modals/color-chart-edit/store/reducer.js
@@ -1,9 +1,0 @@
-import {makeFormReducer} from '#/main/app/content/form/store/reducer'
-
-import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-edit/store/selectors'
-
-const reducer = makeFormReducer(selectors.STORE_NAME, { new: true })
-
-export {
-  reducer
-}
