+++ src/main/theme/Resources/modules/administration/appearance/modals/color-chart-parameters/components/modal.jsx
@@ -5,16 +5,17 @@ import {trans} from '#/main/app/intl/translation'
 import {Modal} from '#/main/app/overlays/modal/components/modal'
 import {FormData} from '#/main/app/content/form/containers/data'
 
+import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store/selectors'
 import {CALLBACK_BUTTON} from '#/main/app/buttons'
 import {Button} from '#/main/app/action'
 
-import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-edit/store/selectors'
-const ColorChartEditModal = props => {
+const ColorChartParametersModal = props => {
+
   return <Modal
     {...omit(props, 'formData', 'saveEnabled', 'save', 'reset', 'updateProp', 'onSave')}
-    icon="fa fa-fw fa-pencil"
-    title={trans('edit_color_chart', {}, 'appearance')}
-    onEntering={() => props.reset(props.colorChart)}
+    icon="fa fa-fw fa-plus"
+    title={trans('new_color_chart', {}, 'appearance')}
+    onExiting={() => props.reset()}
   >
     <FormData
       name={selectors.STORE_NAME}
@@ -65,5 +66,5 @@ const ColorChartEditModal = props => {
 }
 
 export {
-  ColorChartEditModal
+  ColorChartParametersModal
 }
