+++ src/main/theme/Resources/modules/administration/appearance/modals/color-chart-parameters/containers/modal.jsx
@@ -3,10 +3,10 @@ import {connect} from 'react-redux'
 import {withReducer} from '#/main/app/store/components/withReducer'
 import {actions as formActions, selectors as formSelectors} from '#/main/app/content/form/store'
 
-import {actions, reducer, selectors} from '#/main/theme/administration/appearance/modals/color-chart-creation/store'
-import {ColorChartCreationModal as ColorChartCreationModalComponent} from '#/main/theme/administration/appearance/modals/color-chart-creation/components/modal'
+import {actions, reducer, selectors} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store'
+import {ColorChartParametersModal as ColorChartParametersModalComponent} from '#/main/theme/administration/appearance/modals/color-chart-parameters/components/modal'
 
-const ColorChartCreationModal = withReducer(selectors.STORE_NAME, reducer)(
+const ColorChartParametersModal = withReducer(selectors.STORE_NAME, reducer)(
   connect(
     state => ({
       formData: formSelectors.data(formSelectors.form(state, selectors.STORE_NAME)),
@@ -23,9 +23,10 @@ const ColorChartCreationModal = withReducer(selectors.STORE_NAME, reducer)(
         dispatch(formActions.reset(selectors.STORE_NAME, {}, true))
       }
     })
-  )(ColorChartCreationModalComponent)
+  )(ColorChartParametersModalComponent)
 )
 
 export {
-  ColorChartCreationModal
+  ColorChartParametersModal
 }
+
