+++ src/main/theme/Resources/modules/administration/appearance/modals/color-chart-parameters/index.js
@@ -1,0 +1,11 @@
+import {registry} from '#/main/app/modals/registry'
+
+import {ColorChartParametersModal} from '#/main/theme/administration/appearance/modals/color-chart-parameters/containers/modal'
+
+const MODAL_PARAMETERS_COLOR_CHART = 'MODAL_PARAMETERS_COLOR_CHART'
+
+registry.add(MODAL_PARAMETERS_COLOR_CHART, ColorChartParametersModal)
+
+export {
+  MODAL_PARAMETERS_COLOR_CHART
+}
