+++ src/main/theme/Resources/modules/administration/appearance/modals/color-chart-parameters/store/index.js
@@ -1,0 +1,9 @@
+import {reducer} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store/reducer'
+import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store/selectors'
+import {actions} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store/actions'
+
+export {
+  actions,
+  reducer,
+  selectors
+}
