+++ src/main/theme/Resources/modules/administration/appearance/modals/color-chart-parameters/store/reducer.js
@@ -1,6 +1,6 @@
 import {makeFormReducer} from '#/main/app/content/form/store/reducer'
 
-import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-creation/store/selectors'
+import {selectors} from '#/main/theme/administration/appearance/modals/color-chart-parameters/store/selectors'
 
 const reducer = makeFormReducer(selectors.STORE_NAME, { new: true })
